moved everything to mapscreen
