+++ Concordia_Campus_App/maps/MapScreen.js
@@ -2,8 +2,8 @@ import { duration } from 'moment-timezone';
 import React, { useState, useRef, useEffect } from 'react';
 import { GooglePlacesAutocomplete } from 'react-native-google-places-autocomplete';
 import 'react-native-get-random-values';
-import { View, Text, TextInput, TouchableOpacity, Image, StyleSheet } from 'react-native';
-import MapView, { Polygon, Marker, Polyline } from 'react-native-maps';
+import { View, Text, TouchableOpacity, Image, StyleSheet, FlatList, TextInput, Alert } from 'react-native';
+import MapView, { Polygon, Marker } from 'react-native-maps';
 import styles from './styles/mapScreenStyles'; 
 import buildingsData from './buildingCoordinates.js';
 import BuildingPopup from './BuildingPopup'; 
@@ -12,8 +12,9 @@ import { API_KEY } from '@env';
 import ShuttleBusMarker from './ShuttleBusMarker';
 import { getLocation } from './locationUtils';
 import MapDirections from './MapDirections';
+import { useFocusEffect } from '@react-navigation/native';
 
-const MapScreen = () => {
+const MapScreen = ({route}) => {
   const [campus, setCampus] = useState('SGW');
   const [zoomLevel, setZoomLevel] = useState(0.005); 
   const [selectedBuilding, setSelectedBuilding] = useState(null); 
@@ -23,11 +24,21 @@ const MapScreen = () => {
   const [showBuildingDirections, setShowBuildingDirections] = useState(false);
   const [eta, setEta] = useState(null);
   const [distance, setDistance] = useState(null);
+  const [selectedStartBuilding, setSelectedStartBuilding] = useState(null);
   const [selectedStart, setSelectedStart] = useState(null);
+  const [selectedDestination, setSelectedDestination] = useState(null);
   const [selectedEnd, setSelectedEnd] = useState(null); 
   const [shuttleStop, setShuttleStop] = useState(null);
   const [toggleMapDirections, setToggleMapDirections] = useState(false);
   const [userLocation, setUserLocation] = useState(null);
+  const [currentScreen, setCurrentScreen] = useState('Map');
+  
+
+  const [startQuery, setStartQuery] = useState('');
+  const [destinationQuery, setDestinationQuery] = useState('');
+  const [filteredStartBuildings, setFilteredStartBuildings] = useState([]);
+  const [filteredDestinationBuildings, setFilteredDestinationBuildings] = useState([]);
+  const [lastTappedBuilding, setLastTappedBuilding] = useState(null);
   
   const campusLocations = {
     SGW: {
@@ -47,6 +58,121 @@ const MapScreen = () => {
   const location = campusLocations[campus];
 
 
+  const handleStartSearch = (query) => {
+    setStartQuery(query);
+    if (query.length === 0) {
+      setFilteredStartBuildings([]);
+      return;
+    }
+    const results = buildingsData.buildings.filter((b) =>
+      b.name.toLowerCase().includes(query.toLowerCase())
+    );
+    setFilteredStartBuildings(results);
+  };
+
+  const handleDestinationSearch = (query) => {
+    setDestinationQuery(query);
+    if (query.length === 0) {
+      setFilteredDestinationBuildings([]);
+      return;
+    }
+    const results = buildingsData.buildings.filter((b) =>
+      b.name.toLowerCase().includes(query.toLowerCase())
+    );
+    setFilteredDestinationBuildings(results);
+  };
+
+  const handleSelectBuilding = (building) => {
+    Alert.alert(
+      `Select Building`,
+      `Do you want to set "${building.name}" as Start or Destination?`,
+      [
+        {
+          text: "Start",
+          onPress: () => {
+            setSelectedStartBuilding(building);
+            setStartQuery(building.name);
+          },
+        },
+        {
+          text: "Destination",
+          onPress: () => {
+            setSelectedDestination(building);
+            setDestinationQuery(building.name);
+          },
+        },
+        { text: "Cancel", style: "cancel" },
+      ]
+    );
+  };
+
+  const isPointInPolygon = (point, polygon) => {
+    let x = point.longitude, y = point.latitude;
+    let inside = false;
+    for (let i = 0, j = polygon.length - 1; i < polygon.length; j = i++) {
+      let xi = polygon[i].longitude, yi = polygon[i].latitude;
+      let xj = polygon[j].longitude, yj = polygon[j].latitude;
+      let intersect = ((yi > y) !== (yj > y)) && 
+                      (x < (xj - xi) * (y - yi) / (yj - yi) + xi);
+      if (intersect) inside = !inside;
+    }
+    return inside;
+  };
+
+  const handleUseCurrentLocation = async () => {
+    try {
+      const location = await getLocation();
+      const userPoint = { latitude: location.latitude, longitude: location.longitude };
+      
+      const currentBuilding = buildingsData.buildings.find((building) =>
+        isPointInPolygon(userPoint, building.coordinates)
+      );
+      
+      if (currentBuilding) {
+        setSelectedStartBuilding(currentBuilding);
+        setStartQuery(currentBuilding.name);
+        moveToLocation(userPoint.latitude, userPoint.longitude);
+      } else {
+   
+        Alert.alert(
+          "No building found",
+          "You are not within a known building. Would you like to use your current location as the start point?",
+          [
+            {
+              text: "Yes",
+              onPress: () => {
+                
+                const pseudoBuilding = {
+                  name: "My Current Location",
+             
+                  coordinates: [{ latitude: userPoint.latitude, longitude: userPoint.longitude }],
+                  markerCoord: userPoint,
+                  fillColor: 'orange', 
+                  strokeColor: 'orange',
+                };
+                setSelectedStartBuilding(pseudoBuilding);
+                setStartQuery("My Current Location");
+                moveToLocation(userPoint.latitude, userPoint.longitude);
+              },
+            },
+            { text: "No", style: "cancel" },
+          ]
+        );
+      }
+    } catch (error) {
+      Alert.alert("Error", "Could not retrieve current location.");
+    }
+  };
+
+  useFocusEffect(
+    React.useCallback(() => {
+      if(route.params?.setCurrentScreen){
+        setCurrentScreen(route.params.setCurrentScreen);
+      }
+      console.log('Current Screen:', route.params?.setCurrentScreen);
+    }, [route.params?.setCurrentScreen])
+  );
+
   useEffect(() => {
     let interval;
       const fetchUserLocation = async () => {
@@ -62,7 +188,7 @@ const MapScreen = () => {
       return () => {
         if(interval) clearInterval(interval);
       };
-    }, [toggleMapDirections,shuttleStop]);
+    }, [toggleMapDirections,shuttleStop,currentScreen]);
 
 
   async function moveToLocation(latitude, longitude) {
@@ -160,13 +286,15 @@ const MapScreen = () => {
   }, []);
 
   return (
+
     <View style={styles.container}>
+      {currentScreen === 'Map' ? (
       <View style={styles.searchBarContainer}>
         <GooglePlacesAutocomplete
           fetchDetails={true}
           placeholder="Search Building or Class..."
           styles={{
-            textInput: styles.searchBar, 
+            textInput: styles.searchBar,
           }}
           query={{
             key: API_KEY,
@@ -179,30 +307,91 @@ const MapScreen = () => {
               latitude: details?.geometry?.location.lat,
               longitude: details?.geometry?.location.lng,
             });
-              console.log('Selected Marker:', selectedMarker); // Debug marker state
+            console.log('Selected Marker:', selectedMarker); // Debug marker state
           }}
           onFail={(error) => console.log('Error:', error)}
         />
-        {selectedStart && selectedEnd &&(
+        {selectedStart && selectedEnd && (
           <TouchableOpacity
-          style = {styles1.directionsBuildingButton}
-          onPress={handleBuildingDirections}
+            style={styles1.directionsBuildingButton}
+            onPress={handleBuildingDirections}
           >
-          <Image source={require('../assets/location.png')} style={styles1.buttonImage} />
-          <Text style={styles1.directionsBuildingButtonText}>Start</Text>
+            <Image source={require('../assets/location.png')} style={styles1.buttonImage} />
+            <Text style={styles1.directionsBuildingButtonText}>Start</Text>
           </TouchableOpacity>
         )}
-      </View>
-             {selectedStart && selectedEnd && showBuildingDirections && (
-             <MapViewDirections
-             origin={selectedStart}
-             destination={selectedEnd}
-             apikey={API_KEY}
-             strokeWidth={5}
-             strokeColor="blue"
-             onReady={handleDirections}
-             />
-        )}
+      </View>) :     
+      <View style={styles.searchContainer}>
+        <View style={styles.searchBarWrapper}>
+          <Text style={styles.label}>Start:</Text>
+          <TextInput
+            style={[styles.searchBar, { backgroundColor: '#800000', color: '#FFFFFF' }]}
+            placeholder="Select Start Building..."
+            placeholderTextColor="#FFFFFF"
+            value={startQuery}
+            onChangeText={handleStartSearch}
+            onFocus={() => setFilteredStartBuildings([])}
+          />
+          {filteredStartBuildings.length > 0 && (
+            <FlatList
+              data={filteredStartBuildings}
+              keyExtractor={(item) => item.name}
+              renderItem={({ item }) => (
+                <TouchableOpacity onPress={() => {
+                  setSelectedStart(item);
+                  setStartQuery(item.name);
+                  setFilteredStartBuildings([]);
+                }}>
+                  <Text style={styles.searchResultItem}>{item.name}</Text>
+                </TouchableOpacity>
+              )}
+            />
+          )}
+          {/* New button for using current location */}
+          <TouchableOpacity style={styles.useLocationButton} onPress={handleUseCurrentLocation}>
+            <Text style={styles.useLocationButtonText}>Use My Current Location</Text>
+          </TouchableOpacity>
+        </View>
+
+        <View style={styles.searchBarWrapper}>
+          <Text style={styles.label}>Destination:</Text>
+          <TextInput
+            style={[styles.searchBar, { backgroundColor: '#800000', color: '#FFFFFF' }]}
+            placeholder="Select Destination Building..."
+            placeholderTextColor="#FFFFFF"
+            value={destinationQuery}
+            onChangeText={handleDestinationSearch}
+            onFocus={() => setFilteredDestinationBuildings([])}
+          />
+          {filteredDestinationBuildings.length > 0 && (
+            <FlatList
+              data={filteredDestinationBuildings}
+              keyExtractor={(item) => item.name}
+              renderItem={({ item }) => (
+                <TouchableOpacity onPress={() => {
+                  setSelectedDestination(item);
+                  setDestinationQuery(item.name);
+                  setFilteredDestinationBuildings([]);
+                }}>
+                  <Text style={styles.searchResultItem}>{item.name}</Text>
+                </TouchableOpacity>
+              )}
+            />
+          )}
+        </View>
+      </View>}
+  
+      {selectedStart && selectedEnd && showBuildingDirections && (
+        <MapViewDirections
+          origin={selectedStart}
+          destination={selectedEnd}
+          apikey={API_KEY}
+          strokeWidth={5}
+          strokeColor="blue"
+          onReady={handleDirections}
+        />
+      )}
+  
       <View style={styles.toggleButtonContainer}>
         <TouchableOpacity
           style={styles.toggleButton}
@@ -215,9 +404,9 @@ const MapScreen = () => {
           </Text>
         </TouchableOpacity>
       </View>
-
+  
       <MapView
-        ref = {mapRef}
+        ref={mapRef}
         style={styles.map}
         initialRegion={{
           latitude: location.latitude,
@@ -234,16 +423,13 @@ const MapScreen = () => {
       >
         <Marker coordinate={location} title={location.title} description={location.description} />
         <Marker coordinate={destinationLocation} title={destinationLocation.title} description={destinationLocation.description} />
-
-
-        <ShuttleBusMarker setToggleMapDirections={setToggleMapDirections} setShuttleStop={setShuttleStop}/>
-
+  
+        <ShuttleBusMarker setToggleMapDirections={setToggleMapDirections} setShuttleStop={setShuttleStop} />
+  
         {toggleMapDirections && userLocation && shuttleStop && (
-          <MapDirections 
-            userLocation={userLocation} 
-            destinationLocation={shuttleStop}/>
+          <MapDirections userLocation={userLocation} destinationLocation={shuttleStop} />
         )}
-
+  
         {selectedMarker && (
           <Marker
             coordinate={{
@@ -257,7 +443,7 @@ const MapScreen = () => {
             }}
           />
         )}
-
+  
         {buildingsData.buildings.map((building, index) => (
           <Polygon
             key={index}
@@ -268,6 +454,7 @@ const MapScreen = () => {
             onPress={() => handlePolygonPress(building)} // Handle the polygon press
           />
         ))}
+  
         {showDirections && (
           <MapViewDirections
             origin={location}
@@ -278,20 +465,50 @@ const MapScreen = () => {
             onReady={handleDirections}
           />
         )}
-        {selectedStart && selectedEnd && showBuildingDirections &&(
-             <MapViewDirections
-             origin={selectedStart}
-             destination={selectedEnd}
-             apikey={API_KEY}
-             strokeWidth={5}
-             strokeColor="blue"
-             onReady={handleDirections}
-             />
+  
+        {selectedStart && selectedEnd && showBuildingDirections && (
+          <MapViewDirections
+            origin={selectedStart}
+            destination={selectedEnd}
+            apikey={API_KEY}
+            strokeWidth={5}
+            strokeColor="blue"
+            onReady={handleDirections}
+          />
         )}
+
+        {currentScreen === 'Building Map Directions' ? (
+          <>
+            {buildingsData.buildings.map((building) => {
+              let fillColor = building.fillColor;
+              if (building === selectedStartBuilding) fillColor = 'green';
+              else if (building === selectedDestination) fillColor = 'blue';
+
+              return (
+                <Polygon
+                  key={building.name}
+                  coordinates={building.coordinates}
+                  fillColor={fillColor}
+                  strokeColor={building.strokeColor}
+                  strokeWidth={2}
+                  onPress={() => handleSelectBuilding(building)}
+                />
+              );
+            })}
+
+            {lastTappedBuilding && (
+              <Marker
+                coordinate={lastTappedBuilding.markerCoord}
+                title={lastTappedBuilding.name}
+              />
+            )}
+          </>
+        ) : null}
       </MapView>
+  
       {/* Render the BuildingPopup component with the close handler */}
       <BuildingPopup building={selectedBuilding} onClose={handleClosePopup} />
-
+  
       {/* Zoom in/out buttons */}
       <View style={styles.zoomButtonContainer}>
         <TouchableOpacity style={styles.zoomButton} onPress={handleZoomIn}>
@@ -301,29 +518,28 @@ const MapScreen = () => {
           <Text style={styles.zoomButtonText}>−</Text>
         </TouchableOpacity>
       </View>
+  
+      {currentScreen === 'Map' ? (
       <TouchableOpacity
-         style={styles1.directionsButton}
-         onPress={handleCampusDirections}
- 
+        style={styles1.directionsButton}
+        onPress={handleCampusDirections}
       >
-      <View style={styles1.directionsButton}>
-      <Image 
-      source={require('../assets/arrow.png')}  
-      style={styles1.buttonImage} 
-    />
-    
-    <Text style={styles1.directionsButtonText}>{directionsText}</Text>
-  </View>
-</TouchableOpacity>
-
+        <View style={styles1.directionsButton}>
+          <Image source={require('../assets/arrow.png')} style={styles1.buttonImage} />
+          <Text style={styles1.directionsButtonText}>{directionsText}</Text>
+        </View>
+      </TouchableOpacity>
+      ):null}
+  
       {eta !== null && distance !== null && (
-        <View style={[styles1.routeInfoContainer, { flexDirection: 'row'}]}>
+        <View style={[styles1.routeInfoContainer, { flexDirection: 'row' }]}>
           <Text style={styles1.routeInfoText}>Distance: {Math.round(distance)} km</Text>
           <Text style={styles1.routeInfoText}>      ETA: {Math.round(eta)} min</Text>
         </View>
       )}
     </View>
   );
+  
 };
 
 const styles1 = StyleSheet.create({
